Move _convert_streaming_chunks_to_chat_message to utils file
